planflow_backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
dj-rest-auth==7.0.0
django-allauth==0.59.0
pillow==10.1.0